create ImageService and VideoService to reduce controller complexity
