+++ src/Helper/MailSenderHelper.php
@@ -25,25 +25,29 @@ class MailSenderHelper
 
     public function sendMail(string $type, User $user, array $data)
     {
-        $subject = strtoupper($type);
-        $message = (new TemplatedEmail())
-            ->from(new Address('mailer@snowtricks.com', 'No-reply Snowtricks'))
-            ->to(new Address($user->getEmail(), $user->getUsername()))
-            ->context($data)
-            ->htmlTemplate('email/' . $type . '.html.twig');
-        
-        switch ($subject) {
-            case 'ACCOUNT_CONFIRMATION':
-                $message->subject(self::ACCOUNT_CONFIRMATION);
-                break;
-            case 'PASSWORD_RESET':
-                $message->subject(self::PASSWORD_RESET);
-                break;
-            case 'TRICK_REPORT':
-                $message->subject(self::TRICK_REPORT);
-                break;
+        try {
+            $subject = strtoupper($type);
+            $message = (new TemplatedEmail())
+                ->from(new Address('mailer@snowtricks.com', 'No-reply Snowtricks'))
+                ->to(new Address($user->getEmail(), $user->getUsername()))
+                ->context($data)
+                ->htmlTemplate('email/' . $type . '.html.twig');
+
+            switch ($subject) {
+                case 'ACCOUNT_CONFIRMATION':
+                    $message->subject(self::ACCOUNT_CONFIRMATION);
+                    break;
+                case 'PASSWORD_RESET':
+                    $message->subject(self::PASSWORD_RESET);
+                    break;
+                case 'TRICK_REPORT':
+                    $message->subject(self::TRICK_REPORT);
+                    break;
+            }
+
+            $this->mailer->send($message);
+        } catch (\Exception $exception) {
+            throw $exception;
         }
-            
-        $this->mailer->send($message);
     }
 }
